+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.34] - 2023-12-19
+
+### Fixed
+- Operation id wasn't used as function name in JS client.
+- Lack of space before base url in Markdown.
+
+### Added
+- Anchors in Markdown operation tags.
+- Dependencies updated.
+
 ## [0.1.33] - 2023-08-04
 
 ### Added